# Intelligent-System-Architect/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: number;
  username: string;
  email: string;
  createdAt: string;
}

export interface SignupInput {
  /** @minLength 3 */
  username: string;
  email: string;
  /** @minLength 8 */
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface ApiError {
  error: string;
}

export interface MLAnalyzeInput {
  /** Raw CSV string content */
  csvData: string;
  /**
     * Column name to predict (optional)
     * @nullable
     */
  targetColumn?: string | null;
  /** Display name for the dataset */
  datasetName?: string;
}

export interface MLTopValue {
  value: string;
  count: number;
}

export interface MLColumnStat {
  name: string;
  type: string;
  missing: number;
  unique: number;
  /** @nullable */
  min?: number | null;
  /** @nullable */
  max?: number | null;
  /** @nullable */
  mean?: number | null;
  /** @nullable */
  std?: number | null;
  topValues: MLTopValue[];
}

export interface MLModelResult {
  rank: number;
  paradigm: string;
  category: string;
  color: string;
  score: number;
  /** @nullable */
  accuracy?: number | null;
  /** @nullable */
  precision?: number | null;
  /** @nullable */
  recall?: number | null;
  /** @nullable */
  f1?: number | null;
  /** @nullable */
  rmse?: number | null;
  /** @nullable */
  r2?: number | null;
  reasons: string[];
  isImplemented: boolean;
  maturity: string;
}

export interface MLPrediction {
  index: number;
  actual: string;
  predicted: string;
  correct: boolean;
}

export interface MLAnalysisResult {
  datasetName: string;
  rows: number;
  columns: number;
  taskType: string;
  /** @nullable */
  targetColumn: string | null;
  columnStats: MLColumnStat[];
  recommendations: MLModelResult[];
  predictions: MLPrediction[];
  overallBestModel: string;
  summary: string;
  /** @nullable */
  trainAccuracy?: number | null;
  /** @nullable */
  testAccuracy?: number | null;
}