# Intelligent-System-Architect/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: ml
    description: Machine learning operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/signup:
    post:
      operationId: authSignup
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Email or username already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /auth/login:
    post:
      operationId: authLogin
      tags: [auth]
      summary: Log in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /auth/logout:
    post:
      operationId: authLogout
      tags: [auth]
      summary: Log out current user
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /auth/me:
    get:
      operationId: getAuthMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /ml/analyze:
    post:
      operationId: mlAnalyze
      tags: [ml]
      summary: Analyze a dataset and get model recommendations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MLAnalyzeInput"
      responses:
        "200":
          description: Analysis result with model recommendations and predictions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MLAnalysisResult"
        "400":
          description: Invalid CSV or dataset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    AuthUser:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        createdAt:
          type: string
      required: [id, username, email, createdAt]

    SignupInput:
      type: object
      properties:
        username:
          type: string
          minLength: 3
        email:
          type: string
        password:
          type: string
          minLength: 8
      required: [username, email, password]

    LoginInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required: [success]

    ApiError:
      type: object
      properties:
        error:
          type: string
      required: [error]

    MLAnalyzeInput:
      type: object
      properties:
        csvData:
          type: string
          description: Raw CSV string content
        targetColumn:
          type: ["string", "null"]
          description: Column name to predict (optional)
        datasetName:
          type: string
          description: Display name for the dataset
      required: [csvData]

    MLTopValue:
      type: object
      properties:
        value:
          type: string
        count:
          type: integer
      required: [value, count]

    MLColumnStat:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        missing:
          type: integer
        unique:
          type: integer
        min:
          type: ["number", "null"]
        max:
          type: ["number", "null"]
        mean:
          type: ["number", "null"]
        std:
          type: ["number", "null"]
        topValues:
          type: array
          items:
            $ref: "#/components/schemas/MLTopValue"
      required: [name, type, missing, unique, topValues]

    MLModelResult:
      type: object
      properties:
        rank:
          type: integer
        paradigm:
          type: string
        category:
          type: string
        color:
          type: string
        score:
          type: number
        accuracy:
          type: ["number", "null"]
        precision:
          type: ["number", "null"]
        recall:
          type: ["number", "null"]
        f1:
          type: ["number", "null"]
        rmse:
          type: ["number", "null"]
        r2:
          type: ["number", "null"]
        reasons:
          type: array
          items:
            type: string
        isImplemented:
          type: boolean
        maturity:
          type: string
      required: [rank, paradigm, category, color, score, reasons, isImplemented, maturity]

    MLPrediction:
      type: object
      properties:
        index:
          type: integer
        actual:
          type: string
        predicted:
          type: string
        correct:
          type: boolean
      required: [index, actual, predicted, correct]

    MLAnalysisResult:
      type: object
      properties:
        datasetName:
          type: string
        rows:
          type: integer
        columns:
          type: integer
        taskType:
          type: string
        targetColumn:
          type: ["string", "null"]
        columnStats:
          type: array
          items:
            $ref: "#/components/schemas/MLColumnStat"
        recommendations:
          type: array
          items:
            $ref: "#/components/schemas/MLModelResult"
        predictions:
          type: array
          items:
            $ref: "#/components/schemas/MLPrediction"
        overallBestModel:
          type: string
        summary:
          type: string
        trainAccuracy:
          type: ["number", "null"]
        testAccuracy:
          type: ["number", "null"]
      required: [datasetName, rows, columns, taskType, targetColumn, columnStats, recommendations, predictions, overallBestModel, summary]