# Intelligent-System-Architect/lib/api-zod/src/generated/types/authUser.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AuthUser {
  id: number;
  username: string;
  email: string;
  createdAt: string;
}