# Intelligent-System-Architect/lib/api-zod/src/generated/types/mLAnalysisResult.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MLColumnStat } from './mLColumnStat';
import type { MLModelResult } from './mLModelResult';
import type { MLPrediction } from './mLPrediction';

export interface MLAnalysisResult {
  datasetName: string;
  rows: number;
  columns: number;
  taskType: string;
  /** @nullable */
  targetColumn: string | null;
  columnStats: MLColumnStat[];
  recommendations: MLModelResult[];
  predictions: MLPrediction[];
  overallBestModel: string;
  summary: string;
  /** @nullable */
  trainAccuracy?: number | null;
  /** @nullable */
  testAccuracy?: number | null;
}