# Intelligent-System-Architect/lib/api-zod/src/generated/types/mLAnalyzeInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MLAnalyzeInput {
  /** Raw CSV string content */
  csvData: string;
  /**
     * Column name to predict (optional)
     * @nullable
     */
  targetColumn?: string | null;
  /** Display name for the dataset */
  datasetName?: string;
}