# Intelligent-System-Architect/lib/api-zod/src/generated/types/mLColumnStat.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MLTopValue } from './mLTopValue';

export interface MLColumnStat {
  name: string;
  type: string;
  missing: number;
  unique: number;
  /** @nullable */
  min?: number | null;
  /** @nullable */
  max?: number | null;
  /** @nullable */
  mean?: number | null;
  /** @nullable */
  std?: number | null;
  topValues: MLTopValue[];
}