# Intelligent-System-Architect/lib/api-zod/src/generated/types/mLPrediction.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MLPrediction {
  index: number;
  actual: string;
  predicted: string;
  correct: boolean;
}